fix: Bump the versions of `libc6` and `libc-bin` (#21)
* fix: Bump the versions of `libc6` and `libc-bin`

The previously used versions are not available anymore. So, bump the
versions to fix the broken Docker build.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* chore: Upgrade aiohttp to version 3.9.0

The CI job flagged vulnerability CVE-2023-49081 in the previously used
version which is fixed in version 3.9.0.

The diff results from running `poetry add aiohttp=3.9.0`.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* chore: Upgrade `transformers` to version 4.36.0

The previously used version exhibits the vulnerabilities `CVE-2023-6730`
and `CVE-2023-7018`. Both are fixed in version 4.36.0.

The diff results from running `poetry add transformers@4.36.0`.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

---------

Signed-off-by: Frank Viernau <frank_viernau@epam.com>